+++ packages/docs/src/stories/Toast.stories.tsx
@@ -1,0 +1,21 @@
+import type { Meta, StoryObj } from '@storybook/react'
+import { Toast, ToastProps } from '@victor-ignite-ui/react'
+
+export default {
+  title: 'Surfaces/Toast',
+  component: Toast,
+  args: {
+    title: 'Agendamento realizado',
+    description: 'Quarta-feira, 23 de novembro às 16h',
+    duration: 10000,
+  },
+  argTypes: {
+    children: {
+      control: {
+        type: null,
+      },
+    },
+  },
+} as Meta<ToastProps>
+
+export const Primary: StoryObj<ToastProps> = {}