+++ packages/docs/src/stories/Tooltip.stories.tsx
@@ -1,0 +1,25 @@
+import type { Meta, StoryObj } from '@storybook/react'
+import { Text, Tooltip, TooltipProps } from '@victor-ignite-ui/react'
+
+export default {
+  title: 'Surfaces/Tooltip',
+  component: Tooltip,
+  args: {
+    title: 'Agendamento realizado',
+    description: 'Quarta-feira, 23 de novembro às 16h',
+    duration: 10000,
+  },
+  argTypes: {
+    children: {
+      control: {
+        type: null,
+      },
+    },
+  },
+} as Meta<TooltipProps>
+
+export const Primary: StoryObj<TooltipProps> = {
+  args: {
+    children: <Text>Hover</Text>,
+  },
+}