+++ packages/react/src/components/Tooltip/styles.ts
@@ -1,0 +1,22 @@
+import * as Tooltip from '@radix-ui/react-tooltip'
+import { styled } from '../../styles'
+
+export const TooltipContainer = styled(Tooltip.Root, {})
+
+export const TooltipContent = styled(Tooltip.Content, {
+  backgroundColor: '$gray900',
+  borderRadius: '$sm',
+  padding: '$3 $4',
+
+  fontFamily: '$inter',
+  fontSize: '$sm',
+  fontWeight: '$medium',
+  lineHeight: '$short',
+  color: '$gray100',
+})
+
+export const TooltipTrigger = styled(Tooltip.Trigger, {
+  backgroundColor: '$gray900',
+  padding: '$3 $4',
+  borderRadius: '$sm',
+})