+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @victor-ignite-ui/react
 
+## 2.1.0
+
+### Minor Changes
+
+- Add a toast and tooltip components
+
 ## 2.0.0
 
 ### Major Changes