+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @victor-ignite-ui/tokens
 
+## 1.2.0
+
+### Minor Changes
+
+- Add a toast and tooltip components
+
 ## 1.1.0
 
 ### Minor Changes